just generate json directly from the root test-cases.yml
